# openapi.yaml
openapi: 3.0.3
info:
  title: Music Stream Order & Delivery System
  description: |
    API for managing music stream orders and deliveries, tracking customer requests,
    stream consumption, and author payments. Designed for performance testing.
    All operations are multi-tenant, scoped by the required `X-Tenant-ID` request header.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development server
...
paths:
...
  /authors:
    get:
      tags: [Authors]
      operationId: listAuthors
      summary: List all authors
      description: Retrieves all authors for the current tenant.
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: List of authors
          content:
            application/json:
              schema:
                type: object
                properties:
                  authors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Author'
                  total:
                    type: integer
                    description: Total number of authors
    post:
      tags: [Authors]
      operationId: createAuthor
      summary: Create a new author
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthorRequest'
            example:
              name: "Hans Zimmer"
              country: "DE"
      responses:
        '201':
          description: Author created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "VALIDATION_ERROR"
                message: "name is required"
  /authors/{author_id}:
    get:
      tags: [Authors]
      operationId: getAuthor
      summary: Get author by ID
      description: Retrieves a single author by ID. Author must belong to the current tenant.
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: author_id
          in: path
          required: true
          description: Unique identifier of the author
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Author found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          description: Author not found (or belongs to different tenant)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "AUTHOR_NOT_FOUND"
                message: "Author with the given id does not exist"

  # ──────────────────────────────────────────────
  # Customers
  # ──────────────────────────────────────────────
...
components:
  parameters:
    TenantId:
      name: X-Tenant-ID
      in: header
      required: true
      description: Tenant identifier. All operations are scoped to this tenant.
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateAuthorRequest:
      type: object
      required: [name, country]
      properties:
        name:
          type: string
          description: Author's full name
          minLength: 1
          maxLength: 255
        country:
          type: string
          description: ISO 3166-1 alpha-2 country code (for future tax calculations)
          minLength: 2
          maxLength: 2
          example: "DE"
...
    Author:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
          description: Tenant this entity belongs to
        name:
          type: string
        country:
          type: string
          description: ISO 3166-1 alpha-2 country code
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
          enum:
            - VALIDATION_ERROR
            - CUSTOMER_NOT_FOUND
            - AUTHOR_NOT_FOUND
            - MUSIC_STREAM_NOT_FOUND
            - ALREADY_ACTIVE
            - ALREADY_INACTIVE
            - CUSTOMER_INACTIVE
            - NO_APPROVED_REQUEST
            - INSUFFICIENT_BALANCE
            - STREAM_REQUEST_NOT_FOUND
            - STREAM_DELIVERY_NOT_FOUND
        message:
          type: string
          description: Human-readable error description